add cf599

+++ template/template.cpp
@@ -5,7 +5,7 @@ namespace skentagon::util{}
 using namespace skentagon::util;
 
 typedef long long ll;
-#define rep(i,n) for( int (i)=0; (i)<(n); ++(i) )
+#define rep(i,n) for( ll (i)=0; (i)<(n); ++(i) )
 #define rpv(d) for(auto& v:d)
 #define co(n) cout << (n) << endl;
 #define cn(n) cin >> n;
